add get_artifacts + fix github ci
