+++ .github/workflows/ci.yml
@@ -14,15 +14,32 @@ jobs:
   build_tfx_addons:
     runs-on: ubuntu-latest
     timeout-minutes: 60
-
+    strategy:
+      # Test for several versions of TFX in parallel
+      matrix:
+        tfx-version: [==0.26.3, ==0.30.2, ==1.0.0rc1, '']
     steps:
     - uses: actions/checkout@v2
     - name: Set up Python 3.7
       uses: actions/setup-python@v2
       with:
         python-version: 3.7
+    - uses: pre-commit/action@v2.0.3
+      name: Run pre-commit checks (pylint/yapf/isort)
+    - name: Cache pip
+      uses: actions/cache@v2
+      with:
+        # This path is specific to Ubuntu
+        path: ~/.cache/pip
+        # Look to see if there is a cache hit for the corresponding setup.py + TFX version 
+        key: ${{ runner.os }}-pip-${{ hashFiles('setup.py') }}-${{ matrix.tfx-version }}
+        restore-keys: |
+          ${{ runner.os }}-pip--${{ hashFiles('setup.py') }}
+          ${{ runner.os }}-pip-
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip wheel
-        pip install -e .[all]
-    - uses: pre-commit/action@v2.0.3
+        pip install -e .[all, test] tfx${{ matrix.tfx-version }}
+    - name: Run tests
+      run: pytest
+
