+++ .gitignore
@@ -2,3 +2,4 @@ env
 tfx_addons.egg-info
 __pycache__
 .vscode
+.pytest_cache
