+++ .pylintrc
@@ -12,17 +12,24 @@ profile=no
 
 # Add files or directories to the denylist. They should be base names, not
 # paths.
-ignore=CVS
+ignore=proto
 
 # Pickle collected data for later comparisons.
-persistent=yes
+persistent=no
 
 # List of plugins (as comma separated values of python modules names) to load,
 # usually to register additional checkers.
-load-plugins=pylint.extensions.docparams
-accept-no-param-doc=no
+load-plugins=
+
+# Use multiple processes to speed up Pylint.
+jobs=4
 
 [MESSAGES CONTROL]
+
+# Enable the message, report, category or checker with the given id(s). You can
+# either give multiple identifier separated by comma (,) or put this option
+# multiple time. See also the "--disable" option for examples.
+enable=indexing-exception,old-raise-syntax
 
 # Disable the message, report, category or checker with the given id(s). You
 # can either give multiple identifiers separated by comma (,) or put this
@@ -33,19 +40,9 @@ accept-no-param-doc=no
 # --enable=similarities". If you want to run only the classes checker, but have
 # no Warning level messages displayed, use"--disable=all --enable=classes
 # --disable=W"
-disable=design,similarities,no-self-use,attribute-defined-outside-init,locally-disabled,star-args,pointless-except,bad-option-value,global-statement,fixme,suppressed-message,useless-suppression,locally-enabled,no-member,no-name-in-module,import-error,unsubscriptable-object,unbalanced-tuple-unpacking,undefined-variable,not-context-manager,invalid-sequence-index,W,R,C,I
+# TODO(b/154672575): Enable wrong-import-order.
+disable=design,similarities,no-self-use,attribute-defined-outside-init,locally-disabled,star-args,pointless-except,bad-option-value,global-statement,fixme,suppressed-message,useless-suppression,locally-enabled,no-member,no-name-in-module,import-error,unsubscriptable-object,unbalanced-tuple-unpacking,undefined-variable,not-context-manager,useless-object-inheritance,unnecessary-pass,no-else-return,no-else-raise,no-else-continue,wrong-import-order,no-value-for-parameter,misplaced-comparison-constant,no-init,abstract-method,unpacking-non-sequence
 
-# Enable the message, report, category or checker with the given id(s). You can
-# either give multiple identifier separated by comma (,) or put this option
-# multiple time. See also the "--disable" option for examples.
-# W0311 bad-indentation
-# W0312 mixed-indentation
-# C0330 bad-continuation
-# C0301 line-too-long
-# C0326 bad-whitespace
-# W0611 unused-import
-# W0622 redefined-builtin
-enable=indexing-exception,old-raise-syntax,W0311,W0312,C0330,C0301,C0326,W0611,W0622
 
 # Set the cache size for astng objects.
 cache-size=500
@@ -146,7 +143,7 @@ class-rgx=^_?[A-Z][a-zA-Z0-9]*$
 function-rgx=^(?:(?P<camel_case>_?[A-Z][a-zA-Z0-9]*)|(?P<snake_case>_?[a-z][a-z0-9_]*))$
 
 # Regular expression which should only match correct method names
-method-rgx=^(?:(?P<exempt>__[a-z0-9_]+__|next)|(?P<camel_case>_{0,2}[A-Z][a-zA-Z0-9]*)|(?P<snake_case>_{0,2}[a-z][a-z0-9_]*))$
+method-rgx=(?x)^(?:(?P<exempt>_[a-z0-9_]+__|Do|runTest|setUp|tearDown|setUpTestCase|tearDownTestCase|setupSelf|tearDownClass|setUpClass|(test|assert|benchmark)_*[A-Z0-9][a-zA-Z0-9_]*|next)|(?P<camel_case>_{0,2}[A-Z][a-zA-Z0-9_]*)|(?P<snake_case>_{0,2}[a-z][a-z0-9_]*))$
 
 # Regular expression which should only match correct instance attribute names
 attr-rgx=^_{0,2}[a-z][a-z0-9_]*$
@@ -173,7 +170,7 @@ bad-names=
 
 # Regular expression which should only match function or class names that do
 # not require a docstring.
-no-docstring-rgx=(__.*__|main)
+no-docstring-rgx=(_.*|main|test.*|.*test|.*Test|setUp|tearDown|setUpClass|tearDownClass)$
 
 # Minimum line length for functions/classes that require docstrings, shorter
 # ones are exempt.
@@ -189,14 +186,15 @@ max-line-length=80
 ignore-long-lines=(?x)
   (^\s*(import|from)\s
    |\$Id:\s\/\/depot\/.+#\d+\s\$
-   |^[a-zA-Z_][a-zA-Z0-9_]*\s*=\s*("[^"]\S+"|'[^']\S+')
+   |^\s*[a-zA-Z_][a-zA-Z0-9_]*\s*=\s*(".+"|'.+')
    |^\s*\#\ LINT\.ThenChange
    |^[^#]*\#\ type:\ [a-zA-Z_][a-zA-Z0-9_.,[\] ]*$
    |pylint
    |"""
    |\#
    |lambda
-   |(https?|ftp):)
+   |(https?|ftp):
+   |^\s*[rfu]?'.*'$)
 
 # Allow the body of an if to be on the same line as the test if there is no
 # else.
@@ -275,7 +273,7 @@ max-args=5
 
 # Argument names that match this expression will be ignored. Default to name
 # with leading underscore
-ignored-argument-names=_.*
+ignored-argument-names=^\*{0,2}(_$|unused_|dummy_)
 
 # Maximum number of locals for function / method body
 max-locals=15
@@ -322,7 +320,6 @@ deprecated-members=string.atof,string.atoi,string.atol,string.capitalize,string.
 
 [DOCSTRING]
 
-default-docstring-type=google
 # List of exceptions that do not need to be mentioned in the Raises section of
 # a docstring.
 ignore-exceptions=AssertionError,NotImplementedError,StopIteration,TypeError
