+++ setup.cfg
@@ -1,0 +1,8 @@
+
+[aliases]
+test=pytest
+
+[tool:pytest]
+addopts = --verbose
+python_files = *_test.py
+norecursedirs = projects
