+++ tfx_addons/__init__.py
@@ -1,1 +1,3 @@
+"""Init module for TFX."""
+
 __version__ = "0.0.1dev0"
